requirements.txt
# scraperOfertas dependencies
# pip install -r requirements.txt

# Browser automation
playwright>=1.40.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.25.0

# PostgreSQL
asyncpg>=0.29.0
psycopg2-binary>=2.9.5
SQLAlchemy>=2.0.0
alembic>=1.13.0

# Config and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
tzdata>=2024.1

# Tests
pytest>=8.0.0

# Executable build
pyinstaller>=5.0.0
